+++ src/ResourceManager/ManagedServiceIdentity/Commands.ManagedServiceIdentity.Test/TestController.cs
@@ -9,6 +9,7 @@ using Microsoft.Azure.Test.HttpRecorder;
 using Microsoft.Rest.ClientRuntime.Azure.TestFramework;
 using Microsoft.WindowsAzure.Commands.ScenarioTest;
 using Microsoft.Azure.Management.Resources;
+using Microsoft.Azure.ServiceManagemenet.Common.Models;
 
 namespace Microsoft.Azure.Commands.ManagedServiceIdentity.Test
 {
@@ -28,16 +29,18 @@ namespace Microsoft.Azure.Commands.ManagedServiceIdentity.Test
             _helper = new EnvironmentSetupHelper();
         }
 
-        public void RunPsTest(params string[] scripts)
+        public void RunPsTest(XunitTracingInterceptor logger, params string[] scripts)
         {
             var callingClassType = LegacyTest.TestUtilities.GetCallingClass(2);
             var mockName = LegacyTest.TestUtilities.GetCurrentMethodName(2);
+
+            _helper.TracingInterceptor = logger;
 
             RunPsTestWorkflow(
                 () => scripts,
                 // no custom initializer
                 null,
-                // no custom cleanup 
+                // no custom cleanup
                 null,
                 callingClassType,
                 mockName);
@@ -76,7 +79,6 @@ namespace Microsoft.Azure.Commands.ManagedServiceIdentity.Test
                                         .Last();
                 _helper.SetupModules(AzureModule.AzureResourceManager,
                     _helper.RMProfileModule,
-                    _helper.RMResourceModule,
                     _helper.GetRMModulePath("AzureRm.ManagedServiceIdentity.psd1"),
                     "ScenarioTests\\" + callingClassName + ".ps1",
                     "AzureRM.Resources.ps1");