+++ src/ResourceManager/RecoveryServices/Commands.RecoveryServices/Vault/GetAzureRMRecoveryServicesVaultSettingsFile.cs
@@ -32,7 +32,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices
     /// <summary>
     /// Retrieves Azure Recovery Services Vault Settings File.
     /// </summary>
-    [Cmdlet(VerbsCommon.Get, "AzureRmRecoveryServicesVaultSettingsFile")]
+    [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "RecoveryServicesVaultSettingsFile")]
     [OutputType(typeof(VaultSettingsFilePath))]
     public class GetAzureRmRecoveryServicesVaultSettingsFile : RecoveryServicesCmdletBase
     {
@@ -169,6 +169,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices
             string fileName = this.GenerateFileName();
 
             string filePath = string.IsNullOrEmpty(this.Path) ? Utilities.GetDefaultPath() : this.Path;
+            filePath = ResolveUserPath(filePath);
 
             // Generate file.
             if (RecoveryServicesClient.getVaultAuthType(this.Vault.ResourceGroupName, this.Vault.Name) == 0)
@@ -262,6 +263,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices
         public void GetAzureRMRecoveryServicesVaultBackupCredentials()
         {
             string targetLocation = string.IsNullOrEmpty(this.Path) ? Utilities.GetDefaultPath() : this.Path;
+            targetLocation = ResolveUserPath(targetLocation);
             if (!Directory.Exists(targetLocation))
             {
                 throw new ArgumentException(Resources.VaultCredPathException);