+++ src/ServiceManagement/Common/Commands.ScenarioTest/ServiceManagement/ServiceManagementTests.cs
@@ -57,9 +57,8 @@ namespace Microsoft.WindowsAzure.Commands.ScenarioTest
                 SetupManagementClients();
 
                 List<string> modules = new List<string>();
-                
+
                 modules.Add(Path.Combine(EnvironmentSetupHelper.PackageDirectory,@"ServiceManagement\Azure\Compute\AzurePreview.psd1"));
-                modules.Add(Path.Combine(EnvironmentSetupHelper.PackageDirectory,@"ServiceManagement\Azure\Compute\PIR.psd1"));
                 modules.AddRange(Directory.GetFiles(@"Resources\ServiceManagement".AsAbsoluteLocation(), "*.ps1").ToList());
 
                 helper.SetupEnvironment(AzureModule.AzureServiceManagement);