+++ src/StackAdmin/Azs.Backup.Admin/Module/Azs.Backup.Admin/Generated.PowerShell.Commands/SwaggerPathCommands/New-AzsEncryptionKeyBase64.ps1
@@ -1,0 +1,27 @@
+<#
+Copyright (c) Microsoft Corporation. All rights reserved.
+Licensed under the MIT License. See License.txt in the project root for license information.
+#>
+
+<#
+.SYNOPSIS
+    Generate encryption key for infrastructure backups.
+
+.DESCRIPTION
+    Generate encryption key for infrastructure backups.
+
+.EXAMPLE
+
+    PS C:\>New-AzsEncryptionKeyBase64
+
+    Generate encryption key for infrastructure backups.
+
+#>
+function New-AzsEncryptionKeyBase64
+{
+    $tempEncryptionKeyString = ""
+    foreach($i in 1..64) { $tempEncryptionKeyString += -join ((65..90) + (97..122) | Get-Random | % {[char]$_}) }
+    $tempEncryptionKeyBytes = [System.Text.Encoding]::UTF8.GetBytes($tempEncryptionKeyString)
+    $BackupEncryptionKeyBase64 = [System.Convert]::ToBase64String($tempEncryptionKeyBytes)
+    return $BackupEncryptionKeyBase64
+}