+++ src/ResourceManager/Compute.ManagedService/Commands.Compute.ManagedService/ChangeLog.md
@@ -19,6 +19,12 @@
 -->
 ## Current Release
 
+## Version 0.1.3
+* Updated to the latest version of the Azure ClientRuntime.
+
+## Version 0.1.2
+* Updated help files to include full parameter types and correct input/output types.
+
 ## Version 0.1.1
 * Updated to the latest version of the Azure ClientRuntime
 * Set minimum dependency of module to PowerShell 5.0