+++ src/ResourceManager/Dns/Commands.Dns/ChangeLog.md
@@ -19,6 +19,19 @@
 -->
 ## Current Release
 
+## Version 5.0.4
+* Fixed issue with default resource groups not being set.
+* Updated common runtime assemblies
+
+## Version 5.0.3
+* Fixed issue with default resource groups not being set.
+
+## Version 5.0.2
+* Updated to the latest version of the Azure ClientRuntime.
+
+## Version 5.0.1
+* Updated help files to include full parameter types and correct input/output types.
+
 ## Version 5.0.0
 * Set minimum dependency of module to PowerShell 5.0
 * Introduce multiple breaking changes