+++ src/ResourceManager/Subscription/Commands.Subscription/ChangeLog.md
@@ -19,6 +19,13 @@
 -->
 ## Current Release
 
+## Version 0.2.3
+* Updated to the latest version of the Azure ClientRuntime.
+
+## Version 0.2.2
+* Updated help files to include full parameter types and correct input/output types.
+* Fix bug in example for `New-AzureRmSubscription`
+
 ## Version 0.2.1
 * Set minimum dependency of module to PowerShell 5.0